fixed sign in, authorisation control on urls added

+++ app/scripts/rent/renals-admin/rentalListCtrl.js
@@ -6,16 +6,20 @@ angular.module('CarRentalApp').controller('RentalListCtrl', RentalListCtrl);
 function RentalListCtrl($scope, RentalsService) {
     $scope.rentals = [];
     $scope.rentToDelete = {};
+    $scope.loading = false;
 
     var loadData = function () {
+        $scope.loading = true;
         RentalsService.getAll().then(function (response) {
             $scope.rentals = response.data;
-            console.log($scope.rentals);
+            $scope.loading = false;
+
         }, function (response) {
             if (response.code === 401) {
                 alertify.error("You are not admin!");
             }
             alertify.error("Rest service unavailable");
+            $scope.loading = false;
         });
     };
 
@@ -23,7 +27,7 @@ function RentalListCtrl($scope, RentalsService) {
 
     RentalsService.getAll().then(function (response) {
         $scope.rentals = response.data;
-        console.log($scope.rentals);
+
     }, function (response) {
         if (response.code === 401) {
             alertify.error("You are not admin!");
@@ -33,22 +37,24 @@ function RentalListCtrl($scope, RentalsService) {
 
     $scope.removeRent = function (rent) {
         $scope.rentToDelete = rent;
-        alertify.confirm('Are you sure you want to delete rental?', removeRentSuccess, function () {
-        });
+        alertify.confirm('Are you sure you want to delete rental?', processRentSubmit);
     };
 
-    var removeRentSuccess = function () {
-        RentalsService.delete($scope.rentToDelete.id).then(function (response) {
-            alertify.success("Rent deleted!");
-            loadData();
-            $scope.rentTodelete = {};
-        }, function (response) {
-            if (response.code === 401) {
-                alertify.error("Please sign in");
-            }
-            alertify.error("Rest service unavailable!");
-        });
-
+    var processRentSubmit = function (choose) {
+        if (choose) {
+            RentalsService.delete($scope.rentToDelete.id).then(function (response) {
+                alertify.success("Rent deleted!");
+                loadData();
+                $scope.rentTodelete = {};
+            }, function (response) {
+                if (response.code === 401) {
+                    alertify.error("Please sign in");
+                    return;
+                }
+                alertify.error("Rest service unavailable!");
+            });
+        }
     };
+
 
 }
